requirements.txt
Django>=2.0
dj-database-url
django-extensions
django-filter
django-model-utils
djangorestframework
Markdown
prettyconf
psycopg2-binary
